backend/requirements.txt
# Web Framework
Flask==2.3.3
flask-cors==4.0.0

# Machine Learning & NLP
torch==2.1.1
transformers==4.33.0
openai-whisper==20231214

# Translation
indic-translib==0.1.0

# Audio Processing
librosa==0.10.0
pydub==0.25.1
scipy==1.11.3
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.0.1

# Development & Testing (optional)
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
